requirements.txt
Flask==2.3.2
kornia_rs==0.0.8
kornia_moons==0.2.6
numpy==1.24.4
onnx==1.14.0
onnxruntime==1.15.1
onnxsim==0.4.33